Fully separate interface-dependent main routines from definitions of int main()
